编写基类

+++ service/src/main/java/com/cjt/service/security/impl/RoleServiceImpl.java
@@ -2,15 +2,19 @@ package com.cjt.service.security.impl;
 
 import com.alibaba.fastjson.JSONObject;
 import com.caojiantao.common.util.JsonUtils;
+import com.cjt.dao.security.IMenuDAO;
 import com.cjt.dao.security.IRoleDAO;
 import com.cjt.dao.security.IRoleMenusDAO;
 import com.cjt.entity.dto.RoleDTO;
+import com.cjt.entity.model.security.Menu;
 import com.cjt.entity.model.security.Role;
 import com.cjt.service.security.IRoleService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 
 /**
@@ -21,12 +25,15 @@ public class RoleServiceImpl implements IRoleService {
 
     private final IRoleDAO roleDAO;
 
+    private final IMenuDAO menuDAO;
+
     private final IRoleMenusDAO roleMenusDAO;
 
     @Autowired
-    public RoleServiceImpl(IRoleDAO roleDAO, IRoleMenusDAO roleMenusDAO) {
+    public RoleServiceImpl(IRoleDAO roleDAO, IRoleMenusDAO roleMenusDAO, IMenuDAO menuDAO) {
         this.roleDAO = roleDAO;
         this.roleMenusDAO = roleMenusDAO;
+        this.menuDAO = menuDAO;
     }
 
     @Override
@@ -36,35 +43,35 @@ public class RoleServiceImpl implements IRoleService {
 
     @Override
     public JSONObject listRoleByPage(RoleDTO roleDTO) {
-        List<Role> roles = roleDAO.listRole(roleDTO);
-        int total = roleDAO.countRole();
+        List<Role> roles = roleDAO.getList(roleDTO);
+        int total = roleDAO.getListCount(roleDTO);
         return JsonUtils.toPageData(roles, total);
     }
 
     @Override
     public Role getRoleById(int id) {
-        return roleDAO.getRoleById(id);
+        Role role = roleDAO.getById(id);
+        return role;
     }
 
     @Transactional(rollbackFor = Exception.class)
     @Override
     public boolean saveRole(Role role, List<Integer> menuIds) {
-        roleDAO.saveRole(role);
-        roleMenusDAO.saveRoleMenus(role.getId(), menuIds);
-        return role.getId() > 0;
+        if (role.getId() == null) {
+            roleDAO.insert(role);
+            roleMenusDAO.saveRoleMenus(role.getId(), menuIds);
+            return role.getId() > 0;
+        } else {
+            roleMenusDAO.removeRoleMenus(role.getId());
+            roleMenusDAO.saveRoleMenus(role.getId(), menuIds);
+            return roleDAO.updateById(role) > 0;
+        }
     }
 
     @Transactional(rollbackFor = Exception.class)
     @Override
-    public boolean updateRole(Role role, List<Integer> menuIds) {
-        roleMenusDAO.removeRoleMenus(role.getId());
-        roleMenusDAO.saveRoleMenus(role.getId(), menuIds);
-        return roleDAO.updateRole(role) > 0;
-    }
-
-    @Override
     public boolean removeRole(int id) {
         roleMenusDAO.removeRoleMenus(id);
-        return roleDAO.removeRoleById(id) > 0;
+        return roleDAO.deleteById(id) > 0;
     }
 }
